requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-admin-notification==0.2
django-admin-persian-fonts==0.2
django-cleanup==8.0.0
idna==3.4
install==1.3.5
Pillow==9.5.0
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.3
